add animation to menubar, install framer motion
