+++ package.json
@@ -30,6 +30,7 @@
     "clsx": "^2.1.1",
     "lowlight": "^3.3.0",
     "lucide-react": "^0.508.0",
+    "motion": "^12.10.4",
     "next": "15.3.1",
     "next-auth": "^4.24.11",
     "prisma": "^6.6.0",
